+++ csharp-architectural-patterns/PipesAndFilters/App/Program.cs
@@ -1,6 +1,6 @@
 ﻿// See https://aka.ms/new-console-template for more information
 
-using App.Application;
+using App.Domain;
 
 Console.WriteLine(new SentimentAnalyzerPipe().Analyze("I am happy"));
 Console.WriteLine(new SentimentAnalyzerPipe().Analyze("I am sad"));
