+++ csharp-architectural-patterns/PipesAndFilters/Tests/SentimentTest.cs
@@ -1,4 +1,4 @@
-using App.Application;
+using App.Domain;
 
 namespace Tests;
